Added multi-level play, optional silent mode and initial level choice
The game now plays the original simon levels: 8, 14, 20, and 31 items
in the sequence. The sequence lengths can be changed with the LEV1-4
variables.  Additional levels could be added by following the pattern.

If ASK_SILENT is set, the user can press the green button at power on
to run silently; any other button or waiting a few seconds will play
with sound.  This feature is indicated by a solid glowing green LED
at power-up.

If ASK_LEVEL is set, the user can press a button to select the starting
level during the power-up "roulette" display.  This roulette is longer
and faster than the normal roulette in order to indicate that this feature
is active.

If TEST_LEVELS is set, all sequences are filled with only RED to make
it easier to test that the level sequence code works correctly.