Added optional cheat mode
If ASK_CHEAT is set, the user can press the blue button at power on
to cheat; during the game, any button press will be accepted as
correct.  ASK_CHEAT implies ASK_SILENT (because it is only during
the initial pause to wait for silent mode that the cheat option is
available), and it is possible to cheat in silent mode by first
pressing the blue button (cheat) and then pressing the green button
(silent).

Removed TEST_LEVELS because ASK_CHEAT provides the same benefit of
being able to test through all the levels without memorization.